refactor: State 인터페이스 와 관련된 객체 추상화
