+++ src/main/java/blackjack/model/game/MoneyBetter.java
@@ -2,5 +2,6 @@ package blackjack.model.game;
 
 @FunctionalInterface
 public interface MoneyBetter {
+
     int orderToBet(String name);
 }
